重名网络的文件名，并且能够加载paddle的预训练模型
